Change price to float point in db.

+++ app/models.py
@@ -10,7 +10,7 @@ from sqlalchemy.dialects.mysql import TIME
 class SankMerch(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(256), index=True)
-    price = db.Column(db.Integer, index=True)
+    price = db.Column(db.Float, index=True)
     imageLink = db.Column(db.String(256), index=True)
     description = db.Column(db.String(256), index=True)
     quantity = db.Column(db.Integer, index=True)
